COMP: taking out compiler warnings, variables only used in debug macros are seen as unused when compiled in release mode, use size_t for all line wrapping size variables
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=16708
git-svn-id: http://svn.slicer.org/Slicer4/trunk@16708 3bd1e089-480b-0410-8dfb-8563597acbee

+++ QTModules/Annotation/MRML/vtkMRMLAnnotationTextDisplayNode.cxx
@@ -336,13 +336,13 @@ std::string vtkMRMLAnnotationTextDisplayNode::GetLineWrappedText(std::string inp
       }
     return wrappedText;
     }
-  int spaceLeft = maxCharPerLine;
+  size_t spaceLeft = maxCharPerLine;
   vtkDebugMacro("spaceLeft = " << spaceLeft );
   char *line = (char *)(inputText.c_str());
   char *ptr = strtok(line, " ");
   while (ptr != NULL)
     {
-    int wordWidth = strlen(ptr);
+    size_t wordWidth = strlen(ptr);
     vtkDebugMacro("ptr = '" << ptr << "', len = " << wordWidth << ", spaceLeft = " << spaceLeft << ", wrappedText is currently = \n'" << wrappedText.c_str() << "'" );
     // check if adding this word plus a space goes over the line limit
     if (wordWidth + 1 <= spaceLeft)
